Clippy (Cleanups) (#79)
* Clippy

* AUTO Clippy p1

* More clipping

* More clip

* Auto Clip

* Clippy sated!

+++ src/executor/alter_table/truncate.rs
@@ -19,7 +19,7 @@ pub async fn truncate(storage: &mut StorageInner, table_name: &ObjectName) -> Re
 			.try_fold(storage, |storage, column| async move {
 				if column.is_auto_incremented() {
 					storage
-						.set_increment_value(table_name, column.name.value.as_str(), 1 as i64)
+						.set_increment_value(table_name, column.name.value.as_str(), 1_i64)
 						.await?;
 				}
 				Ok(storage)
